% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicator_columns.R
\name{add_indicator_columns}
\alias{add_indicator_columns}
\title{add in all indicators columns from the "TerrADat" layer in a template GDB or a provided list}
\usage{
add_indicator_columns(
  template,
  source,
  all_indicators,
  prefixes_to_zero = c("AH", "FH", "NumSpp", "Spp")
)
}
\description{
Add indicator columns to terradat from a template
}
