% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_unique.R
\name{batch_unique}
\alias{batch_unique}
\title{A faster way of applying 'unique' to a large dataset. Specify a variable to split data by, each subset of the full data will be examined for uniqueness.}
\usage{
batch_unique(indata, splitby)
}
\description{
A faster way of applying 'unique' to a large dataset. Specify a variable to split data by, each subset of the full data will be examined for uniqueness.
}
