% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_unique.R
\name{better_unique}
\alias{better_unique}
\title{A faster way of applying 'unique' to a large dataset.}
\usage{
better_unique(indata)
}
\description{
A faster way of applying 'unique' to a large dataset.
}
