% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ldc.R
\name{cache_ldc}
\alias{cache_ldc}
\title{Fetch data from the LDC server to act as a local cache}
\usage{
cache_ldc(projectkey, verbose, path_cache)
}
\description{
Fetch data from the LDC server to act as a local cache
}
