% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ldc.R
\name{compare_table_to_public}
\alias{compare_table_to_public}
\alias{compare_dataset_to_public}
\title{Compares a specific table to the LDC either cached or downloaded}
\usage{
compare_table_to_public(
  projectkey,
  path_foringest,
  tablename,
  path_cache = NULL,
  path_out = NULL,
  testmode = FALSE
)

compare_dataset_to_public(
  projectkey,
  path_foringest,
  path_cache = NULL,
  path_out = NULL,
  testmode = FALSE
)
}
\description{
Compares a specific table to the LDC either cached or downloaded

Run compare_table_to_public over all tables in a project key
}
