% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_postgres.R
\name{delete_postgres}
\alias{delete_postgres}
\title{Delete data from the postgres server}
\usage{
delete_postgres(
  key,
  user,
  password,
  by = "projectkey",
  schema = "dimadev",
  host = "128.123.177.184",
  port = 5435,
  dbname = "postgres"
)
}
\description{
Delete data from the postgres server
}
