% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dima_prep.R
\name{dima_prep}
\alias{dima_prep}
\title{Attach PrimaryKeys to DIMA Tables}
\usage{
dima_prep(path_dima, speciesstate, path_out = NULL, remove_invalid_keys = T)
}
\description{
Given path to a DIMA and a designation for SpeciesState, create an R object (list) containing DIMA tables with PrimaryKeys appended
}
