% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract a table from a DIMA}
\usage{
extract_table(data.path, dima, query)
}
\description{
Given path, a filename, and a query, extract a table from a DIMA
}
