% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_postgres.R
\name{fetch_postgres}
\alias{fetch_postgres}
\alias{fetch_projectkeys}
\alias{fetch_DBKeys}
\title{Fetch data for a single table from the postgres database}
\usage{
fetch_postgres(
  tbl,
  projectkey,
  user,
  password,
  schema = "dimadev",
  host = "128.123.177.184",
  port = 5435,
  dbname = "postgres"
)

fetch_projectkeys(
  schema,
  host = "jornada-ldc2.jrn.nmsu.edu",
  port = 5432,
  dbname = "postgres",
  user,
  password
)

fetch_DBKeys(
  schema,
  host = "jornada-ldc2.jrn.nmsu.edu",
  port = 5432,
  dbname = "postgres",
  user,
  password
)
}
\description{
Fetch data for a single table from the postgres database
}
