% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forget_who.R
\name{forget_who}
\alias{forget_who}
\title{Remove observer, recorder, and similar from a table}
\usage{
forget_who(data)
}
\arguments{
\item{data}{The table to edit}
}
\value{
A table without names
}
\description{
Given a table, return it without names of data collectors
}
