% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_header_fromenv_tdonly.R
\name{gather_header_fromenv_terradat}
\alias{gather_header_fromenv_terradat}
\title{Generate header from tblPlots object}
\usage{
gather_header_fromenv_terradat(tblPlots)
}
\value{
A dataframe equivalent to what is produced by gather_header()
}
\description{
Generate a header file from a tblPlots object in the environment. TEMPORARY! Need to fold this in to terradactyl.
}
