% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_DIMA}
\alias{ingest_DIMA}
\alias{translate_coremethods}
\alias{ingest_indicators_gdb}
\alias{new_data_only}
\title{Take data from public postgres server and prepare them for ingestion}
\usage{
ingest_DIMA(
  projectkey,
  path_specieslist,
  path_templatetd,
  doLPI = T,
  doGap = T,
  doSR = T,
  doSS = T,
  doHt = T,
  doRH = T,
  doHF = T,
  doGSP = T,
  user = NULL,
  password = NULL
)

translate_coremethods(
  path_tall,
  path_out,
  path_schema,
  projectkey,
  verbose = F
)

ingest_indicators_gdb(path_terradat, path_out, path_schema, verbose = F)

new_data_only(projectkey)
}
\description{
Take data from public postgres server and prepare them for ingestion
}
