% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_qc.R
\name{check_duplicates}
\alias{check_duplicates}
\alias{find_missing_pkeys}
\alias{check_coordinates}
\title{Check for duplicate rows}
\usage{
check_duplicates(tbl)

find_missing_pkeys(tbl, pkeys)

check_coordinates(tblPlots)
}
\description{
Check for duplicate rows

Check for missing primary keys

Check for missing coordinates
}
