% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_to_code.R
\name{texmod_to_code}
\alias{texmod_to_code}
\alias{texture_to_code}
\title{Convert NRCS texture classes/modifiers to codes}
\usage{
texmod_to_code(tex, keep_unrecognized = F)

texture_to_code(tex, keep_unrecognized = F)
}
\arguments{
\item{tex}{Vector of character strings}

\item{keep_unrecognized}{True/False, default to False. If True, unrecognized 
strings will be retained. If False, unrecognized entries will be replaced 
with NA_character.}
}
\value{
A vector of the original length with identified entries changed to 
codes.
}
\description{
Replace NRCS textures or texture modifiers with codes.
}
\examples{
texture_strings <- c("sandy loam", "CLAY", "lerm")
texmod_strings <- c("gravelly", "STONY", "wet")

texture_to_code(texture_strings)
texture_to_code(texture_strings, keep_unrecognized = T)

texmod_to_code(texmod_strings)
texmod_to_code(texmod_strings, keep_unrecognized = T)
}
