% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_schema.R
\name{translate_schema}
\alias{translate_schema}
\title{Translate between schema given a table of associated columns}
\usage{
translate_schema(data, matrix, fromcol, tocol, verbose)
}
\description{
Given data and a table associating two schema, translate the data into the new schema
}
