% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dima_prep.R
\name{dima_prep}
\alias{dima_prep}
\alias{create_species_list}
\title{Attach PrimaryKeys to DIMA Tables}
\usage{
dima_prep(path_dima, speciesstate, path_out = NULL, remove_invalid_keys = T)

create_species_list(path_dima, path_out, speciesstate)
}
\arguments{
\item{path_dima}{Full file path and name of the DIMA to be processed}

\item{speciesstate}{Code to be attached to the data connecting the species list with input plots. Required for terradactyl functions.}

\item{path_out}{If specified, function saves an rdata object containing the function output, and formatted tblSpecies, to this folder.}

\item{remove_invalid_keys}{If TRUE, data which cannot be connected to a valid primary key is dropped. This is recommended to remove orphan records.}
}
\description{
Given path to a DIMA and a designation for SpeciesState, create an R object (list) containing DIMA tables with PrimaryKeys appended

Given path to a DIMA and a designation for SpeciesState, create a species table ready for use by terradactyl
}
