% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_postgres.R, R/fetch_postgres.R
\name{delete_postgres}
\alias{delete_postgres}
\alias{fetch_postgres}
\alias{fetch_projectkeys}
\title{Delete data from the postgres database based on a projectkey}
\usage{
delete_postgres(
  key,
  by = "projectkey",
  schema = "dimadev",
  host = "jornada-ldc2.jrn.nmsu.edu",
  port = 5432,
  dbname = "postgres",
  user = "dima_get",
  password = "dima@1912!"
)

fetch_postgres(
  tbl,
  projectkey,
  schema = "dimadev",
  host = "jornada-ldc2.jrn.nmsu.edu",
  port = 5432,
  dbname = "postgres",
  user = "dima_get",
  password = "dima@1912!"
)

fetch_projectkeys(
  schema,
  host = "jornada-ldc2.jrn.nmsu.edu",
  port = 5432,
  dbname = "postgres",
  user = "dima_get",
  password = "dima@1912!"
)
}
\description{
Delete data from the postgres database based on a projectkey

Fetch data for a single table from the postgres database
}
