% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forget_who.R
\name{forget_who}
\alias{forget_who}
\title{Remove observer and recorder from a table}
\usage{
forget_who(data)
}
\arguments{
\item{data}{The table to edit}
}
\value{
A table without observer or recorder
}
\description{
Given a table, return it without observer, recorder, data enterer, or data checker
}
