% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R, R/translate_schema.R
\name{ingest_DIMA}
\alias{ingest_DIMA}
\alias{new_data_only}
\alias{fetch_splist_from_tdat}
\alias{ingest_indicators_gdb}
\title{Take data from public postgres server and prepare them for ingestion}
\usage{
ingest_DIMA(
  projectkey,
  path_parent,
  path_specieslist,
  path_templatetd,
  path_schema,
  pgschema = "public",
  doLPI = T,
  doGap = T,
  doSR = T,
  doSS = T,
  doHt = T,
  doRH = T,
  doHF = T,
  doGSP = T,
  user = NULL,
  password = NULL
)

new_data_only(projectkey, path_parent, path_cache)

fetch_splist_from_tdat(path_td, speciesstate)

ingest_indicators_gdb(path_terradat, path_out, path_schema, verbose = F)
}
\description{
Take data from public postgres server and prepare them for ingestion
}
