% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_coremethods}
\alias{ingest_coremethods}
\alias{ingest_indicators_gdb}
\title{Prepare tall data for ingestion into the data commons}
\usage{
ingest_coremethods(path_tall, path_out, path_schema, projectkey, verbose = F)

ingest_indicators_gdb(path_terradat, path_out, path_schema, verbose = F)
}
\description{
Given paths to tall tables and a schema, writes
}
